refactor(cef): split host reducer per Phase-H domain (task #182 PR-F-2)
Splits host reducer/mod.rs (1537 lines) into 6 per-Phase-H-domain
files following the same template as launcher PR-B/C/D/E. Pure
mechanical extraction; zero logic changes.

The domain markers were already there as section headers (// ── H.X
— ...); each section becomes its own file.

## Final layout

| File | Lines | Contents |
|---|---|---|
| reducer/mod.rs | 866 | dispatcher (update), Ctx, helpers (emit_error, BROWSER_PANE_LABEL_SEQ + next_browser_pane_label, types) |
| reducer/tests.rs | 770 | all tests |
| reducer/top_level.rs | 249 | H.6 — EnqueueTopLevelWindow, callback_fired, renderer_terminated, externally_closed + start_next_top_level_if_idle + push_top_level_history |
| reducer/pool.rs | 181 | H.4 — pool spawn/ready/destroyed/promote/drain_all |
| reducer/panes.rs | 170 | H.1 — browser pane create/close/drain/abort/try_register |
| reducer/browsers.rs | 61 | H.2 — register/unregister browser |
| reducer/drag.rs | 51 | H.3 — start_drag, end_drag |
| reducer/quit.rs | 38 | H.5 — begin_drain, confirm_drained |

mod.rs went 1537 → 866 lines (-671). Each domain file is focused on
its Phase H scope.

## Cross-module wiring

- All extracted handlers marked pub(super); only mod.rs's update()
  dispatcher calls them.
- emit_error and BROWSER_PANE_LABEL_SEQ + next_browser_pane_label
  stay in mod.rs as pub(super) (used across multiple domains).
- TOP_LEVEL_CREATION_HISTORY_CAP became pub(crate) so top_level.rs
  can reference it.
- panes.rs's next_browser_pane_label call → super::next_browser_pane_label.
- start_next_top_level_if_idle stays in top_level.rs (only top_level
  callers), pub(super) for callback handlers.

63 host tests pass.

This completes the launcher reducer + host reducer per-domain splits
(task #182 PR-A through PR-F-2). The remaining task #182 work is
PR-G (split client.rs).
